Properly override the optimization schema for tasks
https://github.com/taskcluster/taskgraph/commit/c60445965e81d58a78bd6ad01e406820c2e28f66
changed the order things are getting import in which means we do not get
a chance to override OptimizationSchema before the Task schema gets
compiled. So we now override it manually and recompile the schema. It's
not pretty, but as far as I know we don't really have a choice if we
want to continue using the task transform.