Update to 0.6.4
I'm upping the quality threshold again so there's a bunch of yeetage.
First, this shouldn't be a surprise to anyone but I'm tired of having
these fail my gens, we're saying goodbye to ALTTP and OoT. Yes. They're
both unmaintained and have been for so long, I'm done with them. I will
however be replacing with OoT with SoH which while slow, at least has
the merit of having active devs that seem to care about quality and
doesn't have all of the issues OoT had.

Starcraft 2 is another core-verified game that we're saying goodbye to
with its nice 8% failure rate. I am suspecting it to be the one reason I
struggled so hard to get a gen during the last community AP given the
failure type of those gens matches.
Overcooked 2 also has a 11.% failure rate, so goodbye as well.

Non core games on the list of yeetage:

- metroidprime: 2% failure rate. Unfortunately that's probably a forever
  goodbye because that apworld is also breaking another rule that I had
  and I had decided to ignore it while I could. But it's not getting
  back in in this state.
- archipIDLE (it's gone from core, and I won't be missing it)
- ff12 open world, it's intermittently failing some test that it really
  shouldn't (`AssertionError: 197 != 194 : Final Fantasy 12 Open World
  Item count MUST match the number of locations`) and got the
  unfortunate idea of failing it while I was watching
- FFXIV: 4% failure rate
- papermario: Failing one of the newly added tests that is relevant to
  generation success in multiworlds, `AssertionError: 897 != 895 : Paper
  Mario modified locations count during generate_basic`.
- sonic_rush: 2% failure rate
- tje: 2% failure rate
- wordipelago: 1.6% failure rate

Again, those rates are when combined with a second world that has 100
locations with no logic to remove the chances of failures being due to
restrictive starts. I also excluded failures that are early enough that
they'd be caught by the validator (in gens up to generate_basic)